pyproject.toml
[tool.poetry]
name = "dz-24-1"
version = "0.1.0"
description = ""
authors = ["Alexandr Lebedev <alexandr.lebedev.sp@gmail.com>"]
readme = "README.md"
packages = [{include = "dz_24"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
